requirements.txt
# Note: GPU-enabled torch builds require installation via conda or PyTorch extra index URL
tqdm>=4.64.0
requests>=2.28.0
argparse
Shimmy==2.0.0
gymnasium==1.1.1
dm_control==1.0.28
imageio==2.37.0
numpy>=1.24.0
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
matplotlib
